Issue #8: Implemented questions managing subsystem

+++ src/features/topics/actions/deleteTopic.ts
@@ -1,19 +1,29 @@
 'use server';
 
 import { prisma } from '@/lib/db';
+import { getCurrentUser } from '@/lib/session';
 import { isDev } from '@/constants';
 import { TTopic } from '@/features/topics/types';
 
 export async function deleteTopic(topic: TTopic) {
+  // Check user rights to delete the question...?
+  const user = await getCurrentUser();
+  const userId = user?.id;
+  if (!userId) {
+    throw new Error('Undefined user');
+  }
+  // Check user rights to delete the question...
+  if (userId !== topic.userId || user.role !== 'ADMIN') {
+    throw new Error('Current user not allowed to delete the topic');
+  }
   try {
     if (isDev) {
       // DEBUG: Emulate network delay
       await new Promise((resolve) => setTimeout(resolve, 1000));
     }
-    const topicId = topic.id;
     const removedTopic = await prisma.topic.delete({
       where: {
-        id: topicId,
+        id: topic.id,
       },
     });
     return removedTopic;
